Add Oklo Evo Longtail
Added Oklo Evo LongTail. The frame color is RAL 3001, hex color equivalent seems to be #9b2423.

I tried my best to measure the bike's size and margins relative to the border of image but somehow I can't get the wheels to stand on the axis line. Should I crop the image? How to you measure precisely?